Refactor task manager to separate business and view concerns.
